+++ Adapt.PresentationSamples.Standard/Adapt.PresentationSamples.Standard.csproj
@@ -76,6 +76,9 @@
     <Compile Include="Model\DateTimeModel.cs" />
     <Compile Include="Model\ItemModel.cs" />
     <Compile Include="Model\ItemModelProvider.cs" />
+    <Compile Include="Pages\ListViewPage.xaml.cs">
+      <DependentUpon>ListViewPage.xaml</DependentUpon>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="App.xaml">
@@ -84,6 +87,12 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="MainPage.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Pages\ListViewPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
       <SubType>Designer</SubType>
     </EmbeddedResource>
