+++ Adapt.PresentationSamples.Standard/MainPage.xaml
@@ -1,12 +1,12 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <TabbedPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-            xmlns:controls="clr-namespace:Adapt.Presentation.Controls;assembly=Adapt.Presentation.Standard"
             xmlns:converters="clr-namespace:Adapt.Presentation.Converters;assembly=Adapt.Presentation.Standard"
+             xmlns:controls="clr-namespace:Adapt.Presentation.Controls;assembly=Adapt.Presentation.Standard"
             Title="Samples"
+            xmlns:pages="clr-namespace:Pages"
              x:Class="Adapt.PresentationSamples.MainPage"
-            xmlns:local="clr-namespace:TestXamarinForms.AsyncListView">
-    <!--Pages can be added as references or inline-->
+            >
 
     <TabbedPage.Resources>
         <ResourceDictionary>
@@ -137,46 +137,9 @@
         <Button Margin="4" Text="Display Notification" x:Name="NotificationButton"  />
     </ContentPage>
 
-    <ContentPage Title="ListView" x:Name="ListViewPage" >
-        <Grid x:Name="ListViewPageGrid">
+    <pages:ListViewPage Title="ListView" >
 
-            <Grid.Resources>
-                <ResourceDictionary>
-                    <local:ItemModelProvider x:Key="items" />
-
-                    <DataTemplate x:Key="itemtemplate">
-
-                        <Grid >
-
-                            <Grid.RowDefinitions>
-                                <RowDefinition Height="20" />
-                                <RowDefinition Height="20" />
-                            </Grid.RowDefinitions>
-
-                            <Label Text="{Binding Name}" TextColor="#FF0000EE" VerticalOptions="Center"  />
-                            <Label Text="{Binding Description}" Grid.Row="1"  VerticalOptions="Center" />
-
-                        </Grid>
-
-                    </DataTemplate>
-
-                </ResourceDictionary>
-            </Grid.Resources>
-
-            <Grid.RowDefinitions>
-                <RowDefinition Height="33*" />
-                <RowDefinition Height="33*" />
-                <RowDefinition Height="33*" />
-            </Grid.RowDefinitions>
-
-            <ActivityIndicator x:Name="ListViewActivityIndicator" IsRunning="True" />
-
-            <controls:AdaptListView Grid.Row="1" ItemTemplate="{StaticResource itemtemplate}" x:Name="TheListView" SelectionMode="Multi" Margin="2" SelectedItem="{Binding ItemModel, Mode=TwoWay}" ItemsSource="{StaticResource items}" HorizontalOptions="Center" VerticalOptions="Center" BackgroundColor="#EEEEEE" />
-
-            <controls:AdaptListView Grid.Row="2" ItemTemplate="{StaticResource itemtemplate}" ItemsSource="{Binding Source={Reference TheListView}, Path=SelectedItems}" SelectionMode="Single" Margin="2" SelectedItem="{Binding ItemModel, Mode=TwoWay}"  HorizontalOptions="Center" VerticalOptions="Center" BackgroundColor="#EEEEEE" />
-
-        </Grid>
-    </ContentPage>
+    </pages:ListViewPage>
 
 
 </TabbedPage>
