+++ Adapt.PresentationSamples.Standard/Pages/ListViewPage.xaml
@@ -1,0 +1,50 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:local="clr-namespace:TestXamarinForms.AsyncListView"
+             xmlns:controls="clr-namespace:Adapt.Presentation.Controls;assembly=Adapt.Presentation.Standard"
+             
+             x:Class="Pages.ListViewPage">
+    <ContentPage.Content>
+        
+        <Grid x:Name="ListViewPageGrid">
+
+            <Grid.Resources>
+                <ResourceDictionary>
+                    <local:ItemModelProvider x:Key="items" />
+
+                    <DataTemplate x:Key="itemtemplate">
+
+                        <Grid >
+
+                            <Grid.RowDefinitions>
+                                <RowDefinition Height="20" />
+                                <RowDefinition Height="20" />
+                            </Grid.RowDefinitions>
+
+                            <Label Text="{Binding Name}" TextColor="#FF0000EE" VerticalOptions="Center"  />
+                            <Label Text="{Binding Description}" Grid.Row="1"  VerticalOptions="Center" />
+
+                        </Grid>
+
+                    </DataTemplate>
+
+                </ResourceDictionary>
+            </Grid.Resources>
+
+            <Grid.RowDefinitions>
+                <RowDefinition Height="33*" />
+                <RowDefinition Height="33*" />
+                <RowDefinition Height="33*" />
+            </Grid.RowDefinitions>
+
+            <ActivityIndicator x:Name="ListViewActivityIndicator" IsRunning="True" />
+
+            <controls:AdaptListView Grid.Row="1" ItemTemplate="{StaticResource itemtemplate}" x:Name="TheListView" SelectionMode="Multi" Margin="2" SelectedItem="{Binding ItemModel, Mode=TwoWay}" ItemsSource="{StaticResource items}" HorizontalOptions="Center" VerticalOptions="Center" BackgroundColor="#EEEEEE" />
+
+            <controls:AdaptListView Grid.Row="2" ItemTemplate="{StaticResource itemtemplate}" ItemsSource="{Binding Source={Reference TheListView}, Path=SelectedItems}" SelectionMode="Single" Margin="2" SelectedItem="{Binding ItemModel, Mode=TwoWay}"  HorizontalOptions="Center" VerticalOptions="Center" BackgroundColor="#EEEEEE" />
+
+        </Grid>
+        
+    </ContentPage.Content>
+</ContentPage>
